Add passport-jwt strategy into the api to authenticate requests.

+++ controllers/api/v1/posts_api.js
@@ -21,35 +21,20 @@ module.exports.index = async (req, res) => {
 module.exports.delete = async function(req, res) {
     try {
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id) {
-        post.remove();
-        console.log(`${post} --was deleted`);
+        if(post.user == req.user.id) {
+            post.remove();
+            console.log(`${post} --was deleted`);
 
-        await Comment.deleteMany({post: req.params.id});
-
-        // req.flash('success', 'Post Deleted!');
-
-        // if(req.xhr) {
-        //     return res.status(200).json({
-        //         data: {
-        //             post_id: req.params.id
-        //         },
-        //         message: 'Post Deleted',
-        //         flash: {
-        //             'success': req.flash('success'),
-        //             'error': req.flash('error')
-        //         }
-        //     });
-        // }
-        // return res.redirect('back');
-        return res.status(200).json({
-            message: 'Post Deleted'
-        });
-    }
-        // else {
-        //     req.flash('error', 'Cannot Delete Post!');
-        // }
-    catch(err) {
+            await Comment.deleteMany({post: req.params.id});
+            return res.status(200).json({
+                message: 'Post Deleted'
+            });
+        }else{
+            return res.status(401).json({
+                message: 'You Cannot delete this post!'
+            });
+        }
+    } catch(err) {
         console.log(err);
         return res.status(500).json({
             message: 'Internal Server Error'
